backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0